Changed subtotal to a double in the Bill class.

+++ phase1/src/Bill.java
@@ -4,7 +4,7 @@ public class Bill {
 
   public static void outputBill(ArrayList<Dish> orders) {
     String billText = "";
-    int subtotal = 0;
+    double subtotal = 0;
     for (Dish order : orders) {
       subtotal += order.getPrice();
       billText += order.toString() + System.lineSeparator();
@@ -18,7 +18,7 @@ public class Bill {
     String billText = "";
     billText += "Table: " + table.getName() + ", ";
     billText += "CustomerNumber: " + customerNum + ", ";
-    int subtotal = 0;
+    double subtotal = 0;
     for (Dish order : table.getCustomerOrder(customerNum)) {
       billText += order.getStringForBill() + System.lineSeparator();
       subtotal += order.getPrice();
